refactor: :recycle: Extract toolbar to separate component
