+++ src/renderer/components/Terminal/components/Toolbar/style.ts
@@ -1,0 +1,9 @@
+import { Box, styled } from '@mui/material';
+
+export const TerminalToolbar = styled(Box)(({ theme }) => ({
+  borderTop: `1px solid ${theme.palette.primary.main}`,
+  borderBottom: `1px solid ${theme.palette.primary.main}`,
+  display: 'flex',
+  alignItems: 'center',
+  justifyContent: 'flex-end',
+}));
